refactor!: rename source files coze→coz
Renames coze.js→coz.js, coze_array.js→coz_array.js,
coze.html→coz.html. Updates all import paths, BUILD.sh
output names, and package.json main entry point.
