+++ join.js
@@ -11,16 +11,16 @@
 //
 // For production, use the following:
 // ```
-// esbuild join.js      --bundle --format=esm --minify --sourcemap         --outfile=coze.min.js
+// esbuild join.js      --bundle --format=esm --minify --sourcemap         --outfile=coz.min.js
 // ```
 //
 // For single file, human readable/debuggable Javascript. 
 // ```
-// esbuild join.js      --bundle --format=esm                              --outfile=coze.join.js  
+// esbuild join.js      --bundle --format=esm                              --outfile=coz.join.js  
 // ```
 //
 export * from './canon.js';
 export * from './alg.js';
-export * from './coze.js';
+export * from './coz.js';
 export * from './key.js';
 export * from './cryptokey.js';
