+++ key.js
@@ -4,7 +4,7 @@ import * as Can from './canon.js'; // import as "Can".
 import * as Enum from './alg.js';
 import {
 	isEmpty,
-} from './coze.js';
+} from './coz.js';
 
 export {
 	NewKey,
@@ -83,7 +83,7 @@ async function Valid(privateCozKey) {
 		return false;
 	}
 	try {
-		let Coz = await import('./coze.js');
+		let Coz = await import('./coz.js');
 		let msg = `7AtyaCHO2BAG06z0W1tOQlZFWbhxGgqej4k9-HWP3DE-zshRbrE-69DIfgY704_FDYez7h_rEI1WQVKhv5Hd5Q`;
 		let sig = await Coz.SignPayRaw(msg, privateCozKey);
 		return Coz.VerifyPay(msg, privateCozKey, sig);
@@ -173,7 +173,7 @@ async function Correct(ck) {
 	// If private key, validate by signing and verifying.
 	// `pub` must also be populated, for cryptokey, since we do not have RecalcX().
 	if (!isPrvEmpty && !isPubEmpty) {
-		let Coz = await import('./coze.js');
+		let Coz = await import('./coz.js');
 		let cryptoKey = await CTK.CryptoKey.FromCozKey(ck);
 		let mldBuffer = await Coz.SToArrayBuffer("Test Signing")
 		let sig = await CTK.CryptoKey.SignBuffer(cryptoKey, mldBuffer);
@@ -223,7 +223,7 @@ async function Revoke(cozKey, msg) {
 		throw new Error("CozKey.Revoke: Private key not set.  Cannot sign message");
 	}
 
-	let Coz = await import('./coze.js');
+	let Coz = await import('./coz.js');
 
 	var coz = {};
 	coz.pay = {};
