+++ package.json
@@ -2,7 +2,7 @@
   "name": "coz",
   "version": "0.0.1",
   "description": "Coz - A cryptographic JSON messaging specification designed for human readability.",
-  "main": "coze.js",
+  "main": "coz.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
