+++ standard/join_standard.js
@@ -2,14 +2,14 @@
 // `join_standard.js` instructs esbuild to join all Coz standard files into one
 // file.
 // ```
-// esbuild join_standard.js   --bundle --format=esm --minify --sourcemap --outfile=coze_standard.min.js
+// esbuild join_standard.js   --bundle --format=esm --minify --sourcemap --outfile=coz_standard.min.js
 // ```
 
 // Coz Core
 export * from '../canon.js';
 export * from '../alg.js';
-export * from '../coze.js';
+export * from '../coz.js';
 export * from '../key.js';
 export * from '../cryptokey.js';
 // Coz Standard
-export * from '../standard/coze_array.js';
+export * from '../standard/coz_array.js';
